requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-allauth==0.41.0
django-appconf==1.0.4
django-axes==5.6.0
django-heroku==0.3.1
django-ipware==3.0.1
django-markdownx==3.0.1
django-taggit==1.1.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
idna==2.8
importlib-metadata==3.4.0
Markdown==3.3.3
mysql==0.0.2
mysql-connector-python==8.0.21
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==8.1.0
protobuf==3.13.0
psycopg2==2.8.5
PyMySQL==0.10.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
schedule==0.6.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.0
typing-extensions==3.7.4.3
urllib3==1.25.7
whitenoise==5.1.0
wincertstore==0.2
zipp==3.4.0
